move internal codecs to their own package
This more rigorous organization showed that we were using some
codecs in an unrelated wallet database used with electrum. We should
probably introduce a `InternalCommonCodecs` class that is more or less
immutable (elementary codecs not impacted by versions changes, such
as `txCodec` or `bool8`).

```
wire
 |
 `-- internal
 |       |
 |       `-- channel
 |       |      |
 |       |      `-- legacy
 |       |      |      |
 |       |      |      `-- legacy0
 |       |      |      |      |
 |       |      |      |      `-- LegacyChannelCodecs0
 |       |      |      |
 |       |      |      `-- legacy1
 |       |      |      |      |
 |       |      |      |      `-- LegacyChannelCodecs1
 |       |      |      |
 |       |      |      `-- ChannelTypes
 |       |      |
 |       |      `-- ChannelCodecs
 |       |
 |       `-- CommandCodecs
 |
 `-- others
```
