+++ eclair-core/src/main/scala/fr/acinq/eclair/db/pg/PgChannelsDb.scala
@@ -23,7 +23,7 @@ import fr.acinq.eclair.db.ChannelsDb
 import fr.acinq.eclair.db.DbEventHandler.ChannelEvent
 import fr.acinq.eclair.db.Monitoring.Metrics.withMetrics
 import fr.acinq.eclair.db.pg.PgUtils.DatabaseLock
-import fr.acinq.eclair.wire.ChannelCodecs.stateDataCodec
+import fr.acinq.eclair.wire.internal.ChannelCodecs.stateDataCodec
 import grizzled.slf4j.Logging
 
 import java.sql.Statement
