+++ eclair-core/src/main/scala/fr/acinq/eclair/wire/internal/ChannelCodecs.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package fr.acinq.eclair.wire
+package fr.acinq.eclair.wire.internal
 
 import fr.acinq.bitcoin.DeterministicWallet.{ExtendedPrivateKey, KeyPath}
 import fr.acinq.bitcoin.{OutPoint, Transaction, TxOut}
@@ -25,8 +25,9 @@ import fr.acinq.eclair.transactions.Transactions._
 import fr.acinq.eclair.transactions._
 import fr.acinq.eclair.wire.CommonCodecs._
 import fr.acinq.eclair.wire.LightningMessageCodecs._
-import fr.acinq.eclair.wire.legacy.legacy0.LegacyChannelCodecs0
-import fr.acinq.eclair.wire.legacy.legacy1.LegacyChannelCodecs1
+import fr.acinq.eclair.wire.UpdateMessage
+import fr.acinq.eclair.wire.internal.legacy.legacy0.LegacyChannelCodecs0
+import fr.acinq.eclair.wire.internal.legacy.legacy1.LegacyChannelCodecs1
 import grizzled.slf4j.Logging
 import scodec.codecs._
 import scodec.{Attempt, Codec}
